CAAT_2061 Option to set a license for a particular groupId/artifactId/version.
This should give the ability to override the found (or non existent) license for a particular dependency.  This should be configured in the gradle build, not an external file.

If a dependency is file based, then you have to specify this in the Gradle build file.

Changes:
1. Fix several issues from the first ticket
(With default configuration, with default report directory)
Fix bug in report (we used parent dependencies instead of original in reports)
1.1 Simplifying logic
2. Static typing
3. Support for license text url in xml reports
4. Support for file dependencies
5. Support for overriding dependencies licenses